LPD-65051 Move to internal.template.parser
