+++ modules/apps/change-tracking/change-tracking-web/src/test/java/com/liferay/change/tracking/web/internal/template/parser/CTJournalTransformerListenerTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.change.tracking.web.internal.transformer;
+package com.liferay.change.tracking.web.internal.template.parser;
 
 import com.liferay.change.tracking.constants.CTConstants;
 import com.liferay.petra.string.StringBundler;
